pyproject.toml
[project]
name = "qkernel"
version = "0.1.0"
description = "CLI tool for executing Quarto cells through a Jupyter kernel"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jupyter_client>=8.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0.0",
    "matplotlib>=3.10.8",
    "ipykernel>=7.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-timeout>=2.0.0",
]

[project.scripts]
qkernel = "qkernel.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
timeout = 60
